PcDuino/requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.23
SQLAlchemy==0.9.4
Werkzeug==0.9.4
backports.ssl-match-hostname==3.4.0.2
cssmin==0.2.0
git-remote-helpers==0.1.0
itsdangerous==0.24
nose==1.3.3
pyparsing==2.0.2
pyserial==2.7
python-dateutil==2.2
six==1.6.1
webassets==0.9
wsgiref==0.1.2
